6502: add the networking user space
Also tweak a pair of prototypes

+++ Library/libs/recvfrom.c
@@ -2,7 +2,7 @@
 #include <sys/types.h>
 #include <sys/socket.h>
 
-int recvfrom(int fd, void *buf, size_t len, int flags, struct sockaddr *addr, socklen_t *addrlen)
+ssize_t recvfrom(int fd, void *buf, size_t len, int flags, struct sockaddr *addr, socklen_t *addrlen)
 {
     __uarg_t args[7];
     args[0] = 7;
