+++ Library/libs/sendto.c
@@ -2,7 +2,7 @@
 #include <sys/types.h>
 #include <sys/socket.h>
 
-int sendto(int fd, const void *buf, size_t len, int flags, const struct sockaddr *addr, socklen_t addrlen)
+ssize_t sendto(int fd, const void *buf, size_t len, int flags, const struct sockaddr *addr, socklen_t addrlen)
 {
     __uarg_t args[7];
     args[0] = 6;
